Fixed segfault issue in server.c
Turns out it had everything to do with the way the return values for
exiting a thread work.  You see, the exiting thread returns a void * to
some kind of location in memory that it has, and the function receiving
this value must pass a void ** value (let's call it foo and call the
void * mentioned above bar).  So in the exit we set (*foo) = bar
effectively repointing (*foo) to the location in memory that the exiting
thread has already set up.  Of course this means that the data cannot be
on the exiting thread's stack since that will be unavailible after it
exits.

This is something I need to watch out for in the future.